requirements.txt
dhis2.py==2.*
numpy==1.*
pandas==1.*